Fixes #3
