+++ app/globals.css
@@ -5,9 +5,6 @@
 body {
   font-family: Arial, Helvetica, sans-serif;
 }
-
-
-
 @layer base {
   :root {
     --background: 0 0% 100%;
